# faster HashMap equality check