--- a/test/junit/scala/collection/immutable/HashMapTest.scala
+++ b/test/junit/scala/collection/immutable/HashMapTest.scala
@@ -1,12 +1,14 @@
 package scala.collection.immutable
 
-import org.junit.Assert.assertEquals
+import org.junit.Assert._
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
 
+import scala.tools.testing.AllocationTest
+
 @RunWith(classOf[JUnit4])
-class HashMapTest {
+class HashMapTest extends AllocationTest {
 
   private val computeHashF = {
     HashMap.empty.computeHash _
@@ -15,6 +17,7 @@ class HashMapTest {
   @Test
   def canMergeIdenticalHashMap1sWithNullKvs() {
     def m = new HashMap.HashMap1(1, computeHashF(1), 1, null)
+
     val merged = m.merged(m)(null)
     assertEquals(m, merged)
   }
@@ -22,6 +25,7 @@ class HashMapTest {
   @Test
   def canMergeIdenticalHashMap1sWithNullKvsCustomMerge() {
     def m = new HashMap.HashMap1(1, computeHashF(1), 1, null)
+
     val merged = m.merged(m) {
       case ((k1, v1), (k2, v2)) =>
         (k1, v1 + v2)
@@ -48,7 +52,9 @@ class HashMapTest {
 
   @Test
   def canMergeHashMapCollision1WithCorrectMerege() {
-    case class A(k: Int) { override def hashCode = 0 }
+    case class A(k: Int) {
+      override def hashCode = 0
+    }
     val m1 = HashMap(A(0) -> 2, A(1) -> 2)
     val m2 = HashMap(A(0) -> 1, A(1) -> 1)
     val merged = m1.merged(m2) { case ((k, l), (_, r)) => k -> (l - r) }
@@ -71,5 +77,43 @@ class HashMapTest {
     assert(stillSingleElement.isInstanceOf[HashMap.HashMap1[_, _]])
     val twoElemTrie = stillSingleElement + (PoorlyHashed(2) -> 2)
     assert(twoElemTrie.isInstanceOf[HashMap.HashTrieMap[_, _]])
+  }
+
+  def generate(): HashMap[String, String] = {
+    (1 to 1000).map { i => s"key $i" -> s"value $i" }(scala.collection.breakOut)
+  }
+
+  @Test
+  def nonAllocatingIdentical(): Unit = {
+    val base = generate()
+    assertTrue(nonAllocating {
+      base == base
+    })
+  }
+
+  @Test
+  def nonAllocatingNotShared(): Unit = {
+    val base = generate()
+    val notShared = generate()
+
+    assertTrue(nonAllocating {
+      base == notShared
+    })
+    assertTrue(nonAllocating {
+      notShared == base
+    })
+  }
+
+  @Test
+  def nonAllocatingShared(): Unit = {
+    val base = generate()
+    val shared = (base - base.head._1) + base.head
+
+    assertTrue(nonAllocating {
+      base == shared
+    })
+    assertTrue(nonAllocating {
+      shared == base
+    })
   }
 }