Specialise equality checking in the common case - where we are comparing a `HashMap` with another `HashMap`

Allocation are reduced  - for tests 
Before
```
nonAllocatingShared    - 16136 bytes
nonAllocatingNotShared - 16136 bytes
```
After 
no allocations


benchmarks are afster except one case when the maps are 10K long and contain the same keys and values, but don't have structural sharing

before
```
[info] Benchmark                                             (size)  Mode  Cnt          Score         Error  Units
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared       10  avgt   20        146.773 ▒       0.683  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared      100  avgt   20       2674.760 ▒      66.553  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared     1000  avgt   20      34981.565 ▒     147.489  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared    10000  avgt   20     476650.815 ▒    6520.751  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared   100000  avgt   20    8162619.004 ▒  389834.682  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared  1000000  avgt   20   89039389.932 ▒ 7529599.778  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical             10  avgt   20          3.320 ▒       0.072  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical            100  avgt   20          3.336 ▒       0.155  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical           1000  avgt   20          3.283 ▒       0.009  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical          10000  avgt   20          3.304 ▒       0.045  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical         100000  avgt   20          3.282 ▒       0.008  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical        1000000  avgt   20          3.310 ▒       0.088  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared             10  avgt   20        281.047 ▒       0.868  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared            100  avgt   20       4686.888 ▒     350.649  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared           1000  avgt   20      62615.105 ▒    2637.198  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared          10000  avgt   20     878134.430 ▒   12359.797  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared         100000  avgt   20   14366528.698 ▒  382193.350  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared        1000000  avgt   20  153765590.238 ▒ 8026359.807  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared                10  avgt   20        141.260 ▒       2.140  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared               100  avgt   20       2344.717 ▒     146.374  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared              1000  avgt   20      35268.297 ▒     490.699  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared             10000  avgt   20     469588.756 ▒    7917.860  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared            100000  avgt   20    8300822.380 ▒  344675.154  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared           1000000  avgt   20   84340060.629 ▒ 5215737.960  ns/op
```

after 
```
[info] Benchmark                                             (size)  Mode  Cnt          Score         Error  Units
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared       10  avgt   20         61.066 ▒       0.288  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared      100  avgt   20        280.327 ▒       6.648  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared     1000  avgt   20        333.969 ▒      16.837  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared    10000  avgt   20        410.014 ▒       8.288  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared   100000  avgt   20        653.467 ▒      25.404  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingDifferentShared  1000000  avgt   20        692.037 ▒      19.032  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical             10  avgt   20          3.337 ▒       0.133  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical            100  avgt   20          3.363 ▒       0.111  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical           1000  avgt   20          3.468 ▒       0.171  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical          10000  avgt   20          3.367 ▒       0.140  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical         100000  avgt   20          3.347 ▒       0.179  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingIdentical        1000000  avgt   20          3.288 ▒       0.022  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared             10  avgt   20        167.913 ▒       0.859  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared            100  avgt   20       3488.789 ▒      43.233  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared           1000  avgt   20      61236.023 ▒     350.485  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared          10000  avgt   20    2147060.290 ▒   40854.180  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared         100000  avgt   20   14091224.088 ▒  144209.350  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingNotShared        1000000  avgt   20  155706817.857 ▒ 1366210.091  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared                10  avgt   20         20.921 ▒       0.142  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared               100  avgt   20        280.882 ▒      12.614  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared              1000  avgt   20        296.281 ▒       7.070  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared             10000  avgt   20        447.103 ▒      28.182  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared            100000  avgt   20        254.787 ▒       2.039  ns/op
[info] HashMapEqualsBenchmark.nonAllocatingShared           1000000  avgt   20        690.022 ▒      23.636  ns/op
```